requirements.txt
anthropic>=0.7.7
python-dotenv>=1.0.1
requests>=2.31.0
sops>=1.18
pyyaml>=6.0
